Sync: always use correct connections
In particular, this resolves an issue when target database is
on another host, and source host doesn't have same-named DB.
Before this fix `dbList` (now renamed to `sourceDBList`) was
incorrectly checked against `targetDB`.

To avoid any possible confusion of contexts, `r` is gone and
`sr`/`tr` are always used explicitly.